# swagger.yaml
openapi: 3.0.0
info:
  title: Seed API
  description: Workshop Backend Developer for SDP Team

servers:
  - url: http://localhost:3000
    description: Main production server
tags:
  - name: Users
    description: User registration and authentication endpoints
  - name: Seeds
    description: Endpoints related to seeds

paths:
  /users/register:
    post:
      summary: Register a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: 
                  type: string
                password: 
                  type: string
      responses:
        '201':
          description: Successfully created.
        '400':
          description: All input is required.
        '409':
          description: User already exists.

  /users/login:
    post:
      summary: Authenticate and login a user
      tags:
        - Users
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: 
                  type: string
                password: 
                  type: string
      responses:
        '200':
          description: Successfully logged in.
        '400':
          description: Invalid Credentials.
        '409':
          description: User already exists.

  /seeds:
    get:
      summary: Returns a list of seeds
      tags:
        - Seeds
      security:
        - ApiKeyAuth: [] 
      responses:
        '200':
          description: A JSON array of seeds
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

    post:
      summary: Creates a seed
      tags:
        - Seeds
      security:
        - ApiKeyAuth: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/Seed'
      responses: 
        '201':
          description: Successfully created

  /seeds/{id}:
    put:
      summary: Update a seed by ID
      tags:
        - Seeds
      security:
        - ApiKeyAuth: [] 
      parameters:
        - name: id
          in: path
          description: ID of the seed to update
          required: true
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/Seed'
      responses:
        '200':
          description: Successfully updated the seed
        '404':
          description: Seed not found

    delete:
      summary: Delete a seed by ID
      tags:
        - Seeds
      security:
        - ApiKeyAuth: [] 
      parameters:
        - name: id
          in: path
          description: ID of the seed to delete
          required: true
          type: string
      responses:
        '200':
          description: Seed deleted successfully
        '404':
          description: Seed not found

definitions:
  User:
    type: object
    properties:
      _id:
        type: string
      email:
        type: string
      token:
        type: string

  Seed:
    type: object
    properties:
      Seed_RepDate:
        type: number
      Seed_Year:
        type: number
      Seeds_YearWeek:
        type: number
      Seed_Variety:
        type: string
      Seed_RDCSD:
        type: string
      Seed_Stock2Sale:
        type: number
      Seed_Season:
        type: number
      Seed_Crop_Year:
        type: string

components:
  securitySchemes:
    ApiKeyAuth:        
      type: apiKey
      in: header       
      name: x-access-token 

security:
  - ApiKeyAuth: [] 